seperate app.get to pages folder
