--- index.html
@@ -1,93 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-  <head>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="style.css">
-
-
-    <title>Hire Us!</title>
-  </head>
-
-    <body>
-
-  <div class="header">
-   <h1>This is a Hire me quiz</h1>
-   <ol>
-     <li>add 3 questions</li>
-     <li>make a good css layout</li>
-     <li>add simple choice javasctip</li>
-   </ol>
-  </div>
-
-
-  <div class="queSecHtml section">
-    <div class="queHtml">
-      <img src="img/html5.png">
-      <h4>What of the following is not describing HTML?</h4>
-      <a href="https://html.com/"><p>read more</p></a>
-    </div>
-
-    <div class="ansHtml">
-      <ul>
-        <li class="ansHtmlA" onclick="htmlA()">The language in which most websites are written.
-          <div class="answers right"></div>
-        </li>
-
-        <li class="ansHtmlB">Stands for Hyper Text Markup Language.
-          <div class="answers right"></div>
-        </li>
-        <li class="ansHtmlC">Can have generic styles, followed by rules that style more specific elements on the page.
-          <div class="answers wrong"></div>
-        </li>
-        <li class="ansHtmlD">First created by Tim Berners-Lee, Robert Cailliau, and others starting in 1989.
-          <div class="answers right"></div>
-        </li>
-      </ul>
-    </div>
-  </div>
-  <div class="queSecCss section">
-    <div class="queCss">
-      <img src="img/css3.png">
-      <h4>Which is not a way to apply CSS styles to a web page?</h4>
-      <a href="https://www.goskills.com/Development/Articles/CSS-interview-questions-answers"><p>read more</p></a>
-    </div>
-
-    <div class="ansCss">
-      <ul>
-        <li class="ansCssA">Place CSS tag in body section.</li>
-        <li class="ansCssB">Loading an external CSS file using the "link" tag</li>
-        <li class="ansCssC">Using the inline style attribute on an element</li>
-        <li class="ansCssD">Using a "style" block in the "head" section of your HTML</li>
-      </ul>
-    </div>
-  </div>
-  <div class="queSecJs section">
-      <div class="queJs">
-      <img src="img/javascript.png">
-      <h4>JavaScript variables can be?</h4>
-      <a href="https://www.w3schools.com/whatis/whatis_js.asp"><p>read more</p></a>
-    </div>
-
-    <div class="ansJs">
-      <ul>
-        <li class="queJsA">Numbers</li>
-        <li class="queJsB">Arrays</li>
-        <li class="queJsC">Functions</li>
-        <li class="queJsD">Simple Text</li>
-      </ul>
-    </div>
-  </div>
-
-
-
-
-    <script src="index.js"></script>
-
-    <script type="text/javascript" src="jquery-3.3.1.js"></script>
-    <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous">
-    </script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous">
-    </script>
-  </body>
-</html>
