--- index.js
@@ -1,1 +1,0 @@
-var hello = alert("click on each anwers to get the correct answer");
